fix: keep the MPS/CPU inference path in a single dtype
The device placement work made bf16 weights reachable on non-CUDA devices,
which exposed three dtype mismatches that CUDA autocast had been hiding.
Autocast is a no-op on MPS/CPU, so every one of these is a hard error there
while remaining invisible on CUDA.

- VAE decode: cast the generated latent to float32 before denormalising.
  The VAE is always fp32 (it runs with autocast disabled), so a bf16 latent
  is a mixed-dtype matmul on MPS/CPU.

- Text fusion: match the layer-fusion scalars (``layer_weights`` /
  ``layer_scale``) to the frozen LLM's hidden-state dtype. Multiplying a
  fp32 scalar into bf16 hidden states is a dtype error without autocast.

- ODE sampling: take the compute dtype from the transformer's own parameters
  rather than ``next(self.parameters())``. The latter can resolve to the fp32
  layer-fusion scalars and silently promote the whole latent / ODE state to
  float32, which then mismatches the bf16 DiT.

- ODE time grid: keep the grid in float32 and align the per-step time to the
  DiT dtype inside ``fn``. ``torchdiffeq.odeint`` casts the grid back to
  float32 internally, and the sway-sampling transform ``1 - cos(pi/2 * t)``
  loses precision in bf16 near t=0, collapsing leading steps onto the same
  value and tripping odeint's strictly-monotonic ``t`` assertion.

Behaviour on CUDA is unchanged: autocast already promoted each of these
operations to fp32, and the per-step time is still cast to the DiT dtype
before it reaches the network.

+++ src/auk/infer/infer_auk.py
@@ -490,7 +490,10 @@ class AukInfer:
             # the DiT is done for this generation: hand its VRAM back before decoding
             self._offload_hooks[1].offload()
         # --- VAE decode (on the VAE's own device) ---
-        gen_latent = self.vae_model.denormalize(gen_latent.to(self.vae_device))
+        # The VAE is always fp32; cast the (possibly bf16) generated latent back to float32 so
+        # the decode runs in a single dtype. On CUDA autocast hides this, but MPS/CPU reject a
+        # mixed bf16-latent / fp32-VAE matmul.
+        gen_latent = self.vae_model.denormalize(gen_latent.float().to(self.vae_device))
         gen_latent = gen_latent.permute(0, 2, 1)  # [1, D, T_new]
 
         gen_audio = self.vae_model.inference_from_latents(gen_latent).cpu()
